Update README.md
Updates the readme noting that the progrinder template was an early proof of concept and functions separately from the rest of ayase. As it stands, we are unlikely to return to the progrinder template since it does not handle images, but in the future, it could be reintegrated back into ayase proper by recycling the template.